requirements.txt

pillow
tkinter-tooltip
